Split into two files and start adding doc
