+++ microapps/BasemapGalleryApp/app/src/main/java/com/arcgismaps/toolkit/basemapgalleryapp/ViewModel.kt
@@ -116,3 +116,4 @@ class ViewModel(application: Application) : AndroidViewModel(application) {
         arcGISScene.setBasemap(basemap)
     }
 }
+
